More code cleanup from Brendan's review.
In particular move the DSV separator detection to the test support namespace and make it clear in comments that it isn't robust enough for general use.
