+++ pwiz_tools/Skyline/TestUtil/AbstractUnitTestEx.cs
@@ -17,6 +17,7 @@
  * limitations under the License.
  */
 
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -25,6 +26,7 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using pwiz.Skyline;
 using pwiz.Skyline.Model;
 using pwiz.Skyline.Model.Results;
+using pwiz.Skyline.Properties;
 using pwiz.Skyline.Util.Extensions;
 
 
@@ -128,6 +130,55 @@ namespace pwiz.SkylineTestUtil
             }
             AssertEx.ConvertedSmallMoleculeDocumentIsSimilar(docOriginal, doc, Path.GetDirectoryName(docPath), mode);
             return doc;
+        }
+
+        /// <summary>
+        /// Examine the lines of a DSV file an attempt to determine what kind of delimiter it uses
+        /// N.B. NOT ROBUST ENOUGH FOR GENERAL USE - would likely fail, for example, on data that has
+        /// irregular column counts. But still useful in the test context where we aren't handed random
+        /// data sets from users.
+        /// </summary>
+        /// <param name="lines">lines of the file</param>
+        /// <param name="columnCount">return value: column count</param>
+        /// <returns>the identified delimiter</returns>
+        /// <exception cref="LineColNumberedIoException">thrown when we can't figure it out</exception>
+        public static char DetermineDsvDelimiter(string[] lines, out int columnCount)
+        {
+
+            // If a candidate delimiter yields different column counts line to line, it's probably not the right one.
+            // So parse some distance in to see which delimiters give a consistent column count.
+            // NOTE we do see files like that in the wild, but not in our test suite
+            var countsPerLinePerCandidateDelimiter = new Dictionary<char, List<int>>
+            {
+                { TextUtil.SEPARATOR_CSV, new List<int>()},
+                { TextUtil.SEPARATOR_SPACE, new List<int>()},
+                { TextUtil.SEPARATOR_TSV, new List<int>()},
+                { TextUtil.SEPARATOR_CSV_INTL, new List<int>()}
+            };
+
+            for (var lineNum = 0; lineNum < Math.Min(100, lines.Length); lineNum++)
+            {
+                foreach (var sep in countsPerLinePerCandidateDelimiter.Keys)
+                {
+                    countsPerLinePerCandidateDelimiter[sep].Add((new DsvFileReader(new StringReader(lines[lineNum]), sep)).NumberOfFields);
+                }
+            }
+
+            var likelyCandidates = 
+                countsPerLinePerCandidateDelimiter.Where(kvp => kvp.Value.Distinct().Count() == 1).ToArray();
+            if (likelyCandidates.Length > 0)
+            {
+                // The candidate that yields the highest column count wins
+                var maxColumnCount = likelyCandidates.Max(kvp => kvp.Value[0]);
+                if (likelyCandidates.Count(kvp => Equals(maxColumnCount, kvp.Value[0])) == 1)
+                {
+                    var delimiter = likelyCandidates.First(kvp => Equals(maxColumnCount, kvp.Value[0])).Key;
+                    columnCount = maxColumnCount;
+                    return delimiter;
+                }
+            }
+
+            throw new LineColNumberedIoException(Resources.TextUtil_DeterminDsvSeparator_Unable_to_determine_format_of_delimiter_separated_value_file, 1, 1);
         }
     }
 }
